+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Use `accessDeniedXml` in `appRootStatus` handler
+- Add `LmcEnvironment.clientIP` function
 
 ## 1.0.0 - 2021-11-22
 - Initial implementation